refactor(allCallTimers): move timer to own component
