+++ src/plugins/allCallTimers/index.tsx
@@ -7,12 +7,10 @@
 import { definePluginSettings } from "@api/Settings";
 import ErrorBoundary from "@components/ErrorBoundary";
 import { Devs } from "@utils/constants";
-import { useTimer } from "@utils/react";
 import definePlugin, { OptionType } from "@utils/types";
 import { findStoreLazy } from "@webpack";
-import { Tooltip } from "@webpack/common";
 
-import { EyeIcon } from "./EyeIcon";
+import { Timer } from "./Timer";
 
 const VoiceStateStore = findStoreLazy("VoiceStateStore");
 
@@ -124,46 +122,11 @@ export default definePlugin({
             return;
         }
         const startTime = user.joinTime;
-        return <ErrorBoundary>
-            <this.Timer time={startTime} />
-        </ErrorBoundary>;
+
+        return (
+            <ErrorBoundary>
+                <Timer time={startTime} />
+            </ErrorBoundary>
+        );
     },
-
-    Timer({ time }: { time: number; }) {
-        const timer = useTimer({});
-        const formatted = new Date(Date.now() - time).toISOString().substring(11, 19);
-
-        if (settings.store.alwaysShow) {
-            return <p style={{
-                margin: 0,
-                fontWeight: "bold",
-                letterSpacing: -2,
-                fontFamily: "monospace",
-                fontSize: 12,
-                color: "red",
-                position: "absolute",
-                bottom: 0,
-                right: 0,
-                padding: 2,
-                background: "rgba(0,0,0,.5)",
-                borderRadius: 3
-            }
-            }> {formatted}</p>;
-        } else {
-            // show as a tooltip
-            return (
-                <Tooltip text={formatted}>
-                    {({ onMouseEnter, onMouseLeave }) => (
-                        <div
-                            onMouseEnter={onMouseEnter}
-                            onMouseLeave={onMouseLeave}
-                            role="tooltip"
-                        >
-                            <EyeIcon />
-                        </div>
-                    )}
-                </Tooltip>
-            );
-        }
-    }
 });
